+++ src/main/java/org/thisway/triplog/dto/TripLogBriefInfo.java
@@ -1,6 +1,9 @@
 package org.thisway.triplog.dto;
 
+import org.thisway.triplog.entity.TripLog;
+
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 public record TripLogBriefInfo(
         Long vehicleId,
@@ -8,8 +11,17 @@ public record TripLogBriefInfo(
         LocalDateTime startTime,
         LocalDateTime endTime,
         Integer tripMeter,
-        Double latitude,
-        Double longitude
+        String address
 ) {
-
+    public static TripLogBriefInfo from(TripLog tripLog) {
+        return new TripLogBriefInfo(
+                tripLog.getVehicle().getId(),
+                tripLog.getVehicle().getCarNumber(),
+                tripLog.getStartTime(),
+                tripLog.getEndTime(),
+                tripLog.getTotalTripMeter(),
+                Optional.ofNullable(tripLog.getOnAddr()).orElse("") +
+                        Optional.ofNullable(tripLog.getOnAddrDetail()).orElse("")
+        );
+    }
 }
