+++ src/main/java/org/thisway/triplog/dto/response/VehicleDetailResponse.java
@@ -1,7 +1,9 @@
 package org.thisway.triplog.dto.response;
 
+import org.springframework.data.domain.Page;
 import org.thisway.triplog.dto.CurrentDrivingInfo;
 import org.thisway.triplog.dto.TripLogBriefInfo;
+import org.thisway.triplog.entity.TripLog;
 import org.thisway.vehicle.dto.response.VehicleResponse;
 
 import java.util.List;
@@ -14,9 +16,15 @@ public record VehicleDetailResponse(
     public static VehicleDetailResponse from(
             VehicleResponse vehicleResponse,
             CurrentDrivingInfo currentDrivingInfo,
-            List<TripLogBriefInfo> tripLogBriefInfos
+            List<TripLog> tripLogs
     ) {
-        return new VehicleDetailResponse(vehicleResponse, currentDrivingInfo, tripLogBriefInfos);
+        return new VehicleDetailResponse(
+                vehicleResponse,
+                currentDrivingInfo,
+                tripLogs.stream()
+                        .map(TripLogBriefInfo::from)
+                        .toList()
+        );
     }
 
 }
